Change redis image
Now using library/redis (official dockerhub redis image) because we've got some problem with CLI with centos/redis-32-centos7.

The main problem here is that the official image is made to start as root, it can work without any problem on openshift leaving random uid to be set (we're using a high value port, and we don't make some write in protected directories) - but we **must** find another way, either we will make our own image.